Update Copyright
Also update KOS system headers

+++ src/deemon/compiler/interface/ilexer.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2021 Griefer@Work                                       *
+/* Copyright (c) 2018-2022 Griefer@Work                                       *
  *                                                                            *
  * This software is provided 'as-is', without any express or implied          *
  * warranty. In no event will the authors be held liable for any damages      *
@@ -12,7 +12,7 @@
  *    claim that you wrote the original software. If you use this software    *
  *    in a product, an acknowledgement (see the following) in the product     *
  *    documentation is required:                                              *
- *    Portions Copyright (c) 2018-2021 Griefer@Work                           *
+ *    Portions Copyright (c) 2018-2022 Griefer@Work                           *
  * 2. Altered source versions must be plainly marked as such, and must not be *
  *    misrepresented as being the original software.                          *
  * 3. This notice may not be removed or altered from any source distribution. *
@@ -2894,7 +2894,7 @@ PRIVATE ptrdiff_t
 
 PRIVATE WUNUSED NONNULL((1)) DREF DeeObject *DCALL
 token_text(DeeCompilerWrapperObject *__restrict self) {
-	int error;
+	ptrdiff_t error;
 	struct unicode_printer printer = UNICODE_PRINTER_INIT;
 	COMPILER_BEGIN(self->cw_compiler);
 	error = TPP_PrintToken(&unicode_printer_tppappend, &printer);
@@ -2931,7 +2931,7 @@ token_str(DeeCompilerWrapperObject *__restrict self) {
 
 PRIVATE WUNUSED NONNULL((1)) DREF DeeObject *DCALL
 token_repr(DeeCompilerWrapperObject *__restrict self) {
-	int error;
+	ptrdiff_t error;
 	struct unicode_printer printer = UNICODE_PRINTER_INIT;
 	COMPILER_BEGIN(self->cw_compiler);
 	if unlikely(unicode_printer_putc(&printer, '\"'))
